+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CNVetti CHANGELOG
 
+## HEAD (unreleased yet)
+
+- Also writing out `FORMAT/CN` in segmentation BCF file for `cmd segment`.
+
 ## v0.1.0
 
 - Everything is new.